Moved base configuration builder test
