+++ webtester-core/src/main/java/info/novatec/testit/webtester/config/ConfigurationException.java
@@ -6,10 +6,6 @@ import info.novatec.testit.webtester.WebTesterException;
 @SuppressWarnings("serial")
 public class ConfigurationException extends WebTesterException {
 
-    protected ConfigurationException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
     protected ConfigurationException(String message) {
         super(message);
     }
